(refactor) change name of second page as blog post
