+++ src/views/BlogPostsList.vue
